Working on styling the map a bit, and polishing the front-end layout
