+++ app/src/components/BottomPanel.js
@@ -1,0 +1,12 @@
+import React, { useState, useEffect } from "react";
+
+import './BottomPanel.scss'
+
+const BottomPanel = props => {
+    return(
+        <div className="bottom-panel-container">
+        </div>
+    )
+}
+
+export default BottomPanel
