+++ app/src/components/BottomPanel.scss
@@ -1,0 +1,8 @@
+.bottom-panel-container {
+    align-self: flex-end;
+    height: 80px;
+    background-image: url("../assets/mist-forest-background-back.png");
+    background-size: contain;
+    background-repeat: repeat;
+
+}
