+++ app/src/components/Map.js
@@ -1,61 +1,51 @@
-import React, { useState } from "react";
+import React, { useEffect } from "react";
+import ReactDOM from "react-dom";
+import { scaleLinear } from "d3-scale";
 import axios from "axios";
 
-import "./Map.scss"
+var mapStyles = { position: "relative"}
+// var svgStyles = { position: "absolute", top: 100, left: 500, right: 0, bot", width: '100%'}tom: 0 };  
+var svgStyles = { position: "relative", width: '100%', top: '2.5%'}
 
-class Map extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = {}
-  }
+const Map = ({ width, height, nodes, links }) => {
+  var xScale = scaleLinear()
+    .domain([0, 10])
+    .range([10, width]);
+  var yScale = scaleLinear()
+    .domain([0, 15])
+    .range([50, height]);
 
-  componentDidMount() {
-    const key = localStorage.getItem("key")
-    const auth = `Token ${key}`
-
-    axios.get(
-      `${process.env.REACT_APP_BACKEND_URL}/api/adv/init/`, 
-      {
-        headers: { Authorization: auth },
-      })
-    .then(res => {
-        console.log(res)
-      })
-    .catch(err => {
-      console.log("err", err)
-    })
-  }
-
-  render() {
-
-    let elements = []
-    for (let i = 0; i < 400; i++) {
-        elements.push(1)
-    }
-
-    return (
-      <div className="map-container">
-            <div 
-                className="room abc"
-                style={{gridColumn: 5, gridRow: 5}}   
-            >    
-            </div>
-            <div 
-                className="room abc"
-                style={{gridColumn: 5, gridRow: 6}}   
-            >    
-            </div>
-            <div 
-                className="room abc"
-                style={{gridColumn: 20, gridRow: 20}}   
-            >    
-            </div>
-          {/*elements.map(element => {
-              return <div className="room"></div>
-          })*/}
-      </div>
-    );
-  }
-};
+  return (
+    <div id="map" style={mapStyles}>
+      <svg
+        style={svgStyles}
+        width={width}
+        height={height}
+        viewBox={`0 0 ${width} ${height}`}
+      >
+        {links.map((link, i) => (
+          <line
+            key={i}
+            x1={xScale(link.x1)}
+            x2={xScale(link.x2)}
+            y1={yScale(link.y1)}
+            y2={yScale(link.y2)}
+            strokeWidth={2}
+            stroke={"grey"}
+          />
+        ))}
+        {nodes.map((node, i) => (
+          <circle
+            key={i}
+            cx={xScale(node.x)}
+            cy={yScale(node.y)}
+            r="6"
+            fill={node.done === true ? "cyan" : "grey"}
+          />
+        ))}
+      </svg>
+    </div>
+  );
+}
 
 export default Map;
