+++ app/src/views/Game.scss
@@ -2,27 +2,23 @@
     background-color: #0d202f;
     height: 100vh;
     .content-container {
-        border: 1px solid red;
+        background-color: #163750;
         display: flex;
         height: 70vh;
-        width: 100%;
 
         padding-left: 5%;
         padding-right: 5%;
         .game-container {
             background-color: black;
-            border: 1px solid blue;
-            width: 100%;
+            border-radius: 5px;
+            
+            width: 60%;
         }
         .information-container {
             padding-left: 2%;
-            // width: 30%;
-            border: 1px solid orange;
+            width: 20%;
         }
     }
-
     .bottom-container {
-        border: 1px solid yellow;
-        height: 50px;
     }
 }
